[feat] 검색 결과 컴포넌트 분리
