+++ src/components/Common/ApartSearch/ApartSearchResult.tsx
@@ -1,45 +1,13 @@
 // 아파트 검색 결과 컴포넌트
 
-import infoIcon from "@/assets/Search/infoIcon.svg";
-import locationIcon from "@/assets/Search/locationIcon.svg";
-import useLocationPath from "@/hooks/Map/useLocationPath";
 import { useInfiniteScroll } from "@/hooks/Search/useInfinityScroll";
-import { useApartInfoStore } from "@/store/useApartInfoStore";
-import { useMainInfoStore } from "@/store/useMainInfoStore";
-import { useMarkerStore } from "@/store/useMarkerStore";
 import { useSearchStore } from "@/store/useSearchStore";
-import getApartData from "@/utils/api/getApartData";
 import styles from "./ApartSearchResult.module.scss";
+import ApartSearchResultItem from "./ApartSearchResultItem";
 
 export default function ApartSearchResult() {
   const keyword = useSearchStore((state) => state.keyword);
   const { items, loading, ref, totalCount } = useInfiniteScroll(keyword);
-  const map = useMarkerStore((state) => state.map);
-  const setMainInfo = useMainInfoStore((state) => state.setMainInfo);
-  const setApartInfo = useMainInfoStore((state) => state.setApartInfo);
-  const setIsSlide = useMainInfoStore((state) => state.setIsSlide);
-  const setIsDetailInfo = useApartInfoStore((state) => state.setIsDetailInfo);
-  const locationPath = useLocationPath();
-
-  const selectSearchApartHandler = (latitude: number, longitude: number) => {
-    if (map) {
-      const newLocation = new naver.maps.LatLng(latitude, longitude);
-      map.setCenter(newLocation);
-      map.setZoom(17);
-    }
-  };
-
-  const selectSearchApartInfoHandler = async (aptId: string) => {
-    const data = await getApartData(
-      `${import.meta.env.VITE_LOCAL_API_CALL}/${locationPath}/info?id=${aptId}`
-    );
-
-    setIsSlide(true);
-    setIsDetailInfo(null);
-    setMainInfo("SELECT");
-
-    setApartInfo(data.data);
-  };
 
   return (
     <section className={styles.apartSearchResult}>
@@ -49,48 +17,14 @@ export default function ApartSearchResult() {
       <ul className={styles.searchResultContainer}>
         {items.length > 0
           ? items.map((listData, index) => (
-              <li
-                key={index}
-                className={styles.searchResult}
-                onClick={() => {
-                  selectSearchApartHandler(
-                    listData.latitude,
-                    listData.longitude
-                  );
-                }}
-              >
-                <div className={styles.searchTitle}>
-                  <h2>{listData.aptName}</h2>
-                  <p>{listData.aptAddress}</p>
-                </div>
-                <div className={styles.searchButton}>
-                  <button
-                    onClick={(e) => {
-                      e.stopPropagation();
-                      selectSearchApartInfoHandler(listData.aptId);
-                    }}
-                  >
-                    <span>
-                      <img src={infoIcon} alt="정보 보기 아이콘" />
-                    </span>
-                    <em>정보 보기</em>
-                  </button>
-                  <button>
-                    <span>
-                      <img src={locationIcon} alt="위치 보기 아이콘" />
-                    </span>
-                    <em>위치 보기</em>
-                  </button>
-                </div>
-              </li>
+              <ApartSearchResultItem key={index} listData={listData} />
             ))
           : !loading && (
               <li className={styles.noResults}>
-                <span>?</span>죄송합니다.
-                <br /> 현재 검색하신 아파트에 대한 정보를 찾을 수 없습니다.
+                <span>?</span>죄송합니다. <br />
+                현재 검색하신 아파트에 대한 정보를 찾을 수 없습니다.
               </li>
             )}
-
         <li ref={ref} className={styles.loadingIndicator}>
           {loading && <div className={styles.loadingSpinner}></div>}
         </li>
